+++ src/komponenter/onboarding/onboarding.less
