+++ src/komponenter/onboarding/onboarding.test.tsx
