+++ src/komponenter/onboarding/onboardingFooter.tsx
